+++ R/helpers.R
@@ -1513,6 +1513,28 @@ obtain_sample_correlations <- function(data, n, corr, na.data, verbose, ...)
 
 }
 
+# Compute thresholds ----
+#' @noRd
+# Updated 22.07.2023
+obtain_thresholds <- function(categorical_variable)
+{
+
+  # Obtain cumulative sums from frequency table
+  cumulative_sum <- cumsum(fast_table(categorical_variable))
+
+  # Obtain cumulative length
+  cumsum_length <- length(cumulative_sum)
+
+  # Obtain thresholds
+  return(
+    qnorm(
+      cumulative_sum[-cumsum_length] /
+        cumulative_sum[cumsum_length]
+    )
+  )
+
+}
+
 #%%%%%%%%%%%%%%%%%%%%%%%%%
 # QUIET CALLS & LOADS ----
 #%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -2858,30 +2880,31 @@ cluster_kappa <- function(base, comparison)
 }
 
 #' @noRd
-# Root mean square error (for data) ----
-# Updated 10.02.2024
-data_rmse <- function(predicted, observed){
-  return(sqrt(mean((predicted - observed)^2, na.rm = TRUE)))
-}
+# Continuous Accuracy (for single variable) ----
+# Updated 12.02.2024
+continuous_accuracy <- function(prediction, observed)
+{
 
-#' @noRd
-# R-squared (for data) ----
-# Updated 10.02.2024
-data_r_squared <- function(predicted, observed){
+  # Compute square error
+  square_error <- (prediction - observed)^2
 
+  # Return accuracies
   return(
-    1 - (
-      sum((predicted - observed)^2, na.rm = TRUE) /
-      sum((observed - mean(observed, na.rm = TRUE))^2, na.rm = TRUE)
+    c(
+      R2 = 1 - (
+        sum(square_error, na.rm = TRUE) /
+        sum((observed - mean(observed, na.rm = TRUE))^2, na.rm = TRUE)
+      ),
+      RMSE = sqrt(mean(square_error, na.rm = TRUE))
     )
   )
 
 }
 
 #' @noRd
-# Accuracy (for data) ----
-# Updated 10.02.2024
-data_accuracy <- function(prediction, observed)
+# Categorical Accuracy (for single variable) ----
+# Updated 12.02.2024
+categorical_accuracy <- function(prediction, observed)
 {
 
   # Get maximum categories
@@ -2911,8 +2934,14 @@ data_accuracy <- function(prediction, observed)
   # Get total values
   total_values <- sum(frequency)
 
+  # Get maximum possible distance incorrect
+  max_distance <- pmax(
+    abs(max_categories - category_sequence), # distance from maximum category
+    abs(min(category_sequence) - category_sequence) # distance from minimum category
+  )
+
   # Get minimum possible weight accuracy
-  minimum_weighted <- 0.5^(max_categories - 1) * total_values
+  minimum_weighted <- sum(0.5^max_distance * frequency)
 
   # Get diagonal of table (correct predictions)
   correct <- diag(accuracy_table)